requirements.txt
frozenlist==1.3.1
discord.py
ruamel.yaml
pytz~=2020.1
schedule~=0.6.0
aiohttp
rapptz-asqlite
pdf2image
aiogoogle
git+https://github.com/Rapptz/asqlite.git
psutil